Move IncubationLogger
