+++ subprojects/build-init/src/main/java/org/gradle/buildinit/plugins/internal/maven/PomProjectInitDescriptor.java
@@ -33,7 +33,7 @@ import org.gradle.buildinit.plugins.internal.modifiers.BuildInitTestFramework;
 import org.gradle.buildinit.plugins.internal.modifiers.ComponentType;
 import org.gradle.buildinit.plugins.internal.modifiers.Language;
 import org.gradle.buildinit.plugins.internal.modifiers.ModularizationOption;
-import org.gradle.util.IncubationLogger;
+import org.gradle.util.internal.IncubationLogger;
 import org.gradle.workers.WorkerExecutor;
 import org.gradleinternal.buildinit.plugins.internal.maven.Maven2GradleWorkAction;
 import org.gradleinternal.buildinit.plugins.internal.maven.MavenConversionException;
