+++ subprojects/core/src/test/groovy/org/gradle/testfixtures/ProjectBuilderTest.groovy
@@ -25,7 +25,7 @@ import org.gradle.internal.featurelifecycle.UsageLocationReporter
 import org.gradle.internal.operations.BuildOperationProgressEventEmitter
 import org.gradle.test.fixtures.file.LeaksFileHandles
 import org.gradle.test.fixtures.file.TestNameTestDirectoryProvider
-import org.gradle.util.IncubationLogger
+import org.gradle.util.internal.IncubationLogger
 import org.gradle.util.internal.Resources
 import org.junit.Rule
 import spock.lang.Ignore
