+++ subprojects/internal-integ-testing/src/main/groovy/org/gradle/integtests/fixtures/executer/InProcessGradleExecuter.java
@@ -80,7 +80,7 @@ import org.gradle.tooling.internal.provider.serialization.SerializeMap;
 import org.gradle.util.CollectionUtils;
 import org.gradle.util.GUtil;
 import org.gradle.util.GradleVersion;
-import org.gradle.util.IncubationLogger;
+import org.gradle.util.internal.IncubationLogger;
 import org.hamcrest.Matcher;
 import org.hamcrest.StringDescription;
